Refactor: Extract display-related functions to a separate helper file
This keeps the main logic files short
