+++ src/calendar-model/scroll-view-model.ts
@@ -1,11 +1,11 @@
 import { LineType } from '../components/Page'
 import { Ref } from '../ref'
 import { LeiningGenerator } from './generator'
-import { LeiningAliyah, LeiningRun } from './model-types'
+import { LeiningRun } from './model-types'
 import IntegerIterator from '../integer-iterator.ts'
 import { physicalLocationFromRef } from '../location.ts'
 import { HDate } from '@hebcal/core'
-import { numVersesBetween } from './hebcal-conversions.ts'
+import { containsRef, getLabels } from './display.ts'
 
 /** The current state exposed by the view model. */
 export interface ViewModelState {
@@ -122,92 +122,11 @@ export class ScrollViewModel {
         ...rawLine,
         verses,
         run,
-        labels: this.getLabels(relevantRuns, verses),
+        labels: getLabels(relevantRuns, verses),
       }
     })
     return { lines, run: lines.find((o) => o.run)!.run! }
   }
-
-  /**
-   * Computes the set of עלייה labels to display for a line:
-   *  - If one or (for מפטיר and שביעי) more עליות begin in this line, return them.
-   *  - If an עלייה ends at this line, but a new עלייה or פרשה begins in the next פסוק, return nothing.
-   *  - If an עלייה ends at this line, and a new עלייה does not begin in the vicinity, return `סוף`.
-   * @param runs The run containing the פסוקים, as well as the next run in the scroll.
-   * @param verses The פסוקים that begin in this line, if any.
-   */
-  private getLabels(runs: LeiningRun[], verses: Ref[]): string[] {
-    if (!runs.length || !verses.length) return []
-    const labels: string[] = []
-    const allAliyot = runs.flatMap((r) => r.aliyot)
-    for (const v of verses) {
-      const myAliyot = runs[0].aliyot
-      // First, look for an עלייה in the containing run.
-      const starts = myAliyot.filter((a) => a.index && refEquals(a.start, v))
-      labels.push(...starts.map((a) => aliyahName(a.index, runs[0])))
-      // If we already have labels, we're done.
-      if (starts.length) continue
-
-      const ends = myAliyot.filter((a) => refEquals(a.end, v))
-      // If no עלייה ends here, there is nothing to label.
-      if (!ends.length) continue
-
-      const followUpStarts = allAliyot.some(
-        (a) => numVersesBetween(v, a.start) === 1
-      )
-      // If another עלייה (in this run or another run) starts in the next פסוק,
-      // we don't need to label the end.
-      if (followUpStarts) continue
-      labels.push(`סוף ${aliyahName(ends[0].index, runs[0])}`)
-    }
-    return labels
-  }
-}
-
-const aliyahStrings = [
-  'ראשון',
-  'שני',
-  'שלישי',
-  'רביעי',
-  'חמישי',
-  'ששי',
-  'שביעי',
-]
-
-function aliyahName(index: LeiningAliyah['index'], run: LeiningRun) {
-  if (index === 'Maftir') return 'מפטיר'
-  if (index < 1 || index > aliyahStrings.length) return null
-
-  if (index === 1) return run.leining.date.title
-  if (run.leining.date.title === 'שמחת תורה') {
-    if (index === 6) return `חתן תורה`
-    if (index === 7) return `חתן בראשית`
-  }
-  return aliyahStrings[index - 1]
-}
-
-function containsRef(o: LeiningAliyah | LeiningRun, ref: Ref): boolean {
-  if (isRun(o)) return o.aliyot.some((a) => containsRef(a, ref))
-  return compareRefs(o.start, ref) <= 0 && compareRefs(o.end, ref) >= 0
-}
-
-function compareRefs(a: Ref, b: Ref): -1 | 0 | 1 {
-  if (a.b != b.b) return compare(a.b, b.b)
-  if (a.c != b.c) return compare(a.c, b.c)
-  return compare(a.v, b.v)
-}
-
-function refEquals(a: Ref, b: Ref): boolean {
-  return compareRefs(a, b) === 0
-}
-
-function compare(a: number, b: number): -1 | 0 | 1 {
-  if (a === b) return 0
-  return a < b ? -1 : 1
-}
-
-function isRun(o: object): o is LeiningRun {
-  return 'aliyot' in o
 }
 
 // TODO(#130): Delete this function & type once we unify these types.
